requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pinecone-client==2.2.4
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
google-generativeai==0.3.2
httpx==0.25.2
asyncio==3.4.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
